+++ app/presenters/export/csv/school_presenter_module.rb
@@ -97,15 +97,15 @@ module Export::Csv::SchoolPresenterModule
   alias_method :school_sharepoint_link_with_academy_label, :school_sharepoint_folder
 
   def school_main_contact_name
-    school_or_academy_contact&.name
+    @project.establishment_main_contact&.name
   end
 
   def school_main_contact_email
-    school_or_academy_contact&.email
+    @project.establishment_main_contact&.email
   end
 
   def school_main_contact_role
-    school_or_academy_contact&.title
+    @project.establishment_main_contact&.title
   end
 
   def headteacher_contact_name
@@ -125,9 +125,5 @@ module Export::Csv::SchoolPresenterModule
     return unless @project.key_contacts&.headteacher.present?
 
     @project.key_contacts.headteacher.email
-  end
-
-  private def school_or_academy_contact
-    @contacts_fetcher.school_or_academy_contact
   end
 end