+++ CHANGELOG.md
@@ -10,6 +10,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Requests to the Academies API now includes a User-Agent HTTP Header
 - AOPU becomes SOPU
+- The primary contact for school in the export no longer uses the first contact
+  if one is not set.
+- The primary contact for the school is now labelled in the export.
 
 ### Added